+++ src/managednetworkfabric/azext_managednetworkfabric/tests/latest/test_acl.py
@@ -25,6 +25,7 @@ def call_scenario1(test):
     setup_scenario1(test)
     step_create(test, checks=[])
     step_show(test, checks=[])
+    step_update(test, checks=[])
     step_list_resource_group(test, checks=[])
     step_list_subscription(test, checks=[])
     step_delete(test, checks=[])
@@ -36,7 +37,7 @@ def step_create(test, checks=None):
         checks = []
     test.cmd(
         'az networkfabric acl create --resource-group {rg} --location {location} --resource-name {name} --configuration-type {configuration_type}'
-        ' --default-action {default_action} --dynamic-match-configurations {dynamic_match_configurations}'
+        ' --default-action {default_action} '
         ' --match-configurations {match_configurations}', checks=checks)
 
 def step_show(test, checks=None):
@@ -45,6 +46,15 @@ def step_show(test, checks=None):
         checks = []
     test.cmd(
         'az networkfabric acl show --resource-name {name} --resource-group {rg}')
+
+def step_update(test, checks=None):
+    '''Access Control List update operation'''
+    if checks is None:
+        checks = []
+    test.cmd(
+        'az networkfabric acl update --resource-group {rg} --resource-name {name}'
+        ' --configuration-type "Inline" --default-action {default_action}'
+        ' --match-configurations {updated_match_configurations}', checks=checks)
 
 def step_list_resource_group(test, checks=None):
     '''Access Control List list by resource group operation'''
@@ -78,6 +88,7 @@ class GA_AccessControlListsScenarioTest1(ScenarioTest):
             'configuration_type': CONFIG.get('ACCESS_CONTROL_LIST', 'configuration_type'),
             'default_action': CONFIG.get('ACCESS_CONTROL_LIST', 'default_action'),
             'dynamic_match_configurations': CONFIG.get('ACCESS_CONTROL_LIST', 'dynamic_match_configurations'),
+            'updated_match_configurations': CONFIG.get('ACCESS_CONTROL_LIST', 'updated_match_configurations'),
             'match_configurations': CONFIG.get('ACCESS_CONTROL_LIST', 'match_configurations'),
         })